+++ tests/openxml/wordprocessing/test_level.py
@@ -21,63 +21,52 @@ class LevelTestCase(TestCase):
         return Level.load(root)
 
     def test_ilvl_mapped_to_level_id_attribute(self):
-        xml = b'''
-            <lvl ilvl="100">
-            </lvl>
-        '''
+        xml = '<lvl ilvl="100"></lvl>'
         level = self._load_from_xml(xml)
-        self.assertEqual(level.level_id, "100")
+        self.assertEqual(level.level_id, '100')
 
     def test_starting_position_attribute(self):
-        xml = b'''
-            <lvl>
-                <start val="50" />
-            </lvl>
-        '''
+        xml = '<lvl><start val="50" /></lvl>'
         level = self._load_from_xml(xml)
-        self.assertEqual(level.start, "50")
+        self.assertEqual(level.start, '50')
 
     def test_num_format_attribute(self):
-        xml = b'''
-            <lvl>
-                <numFmt val="decimal" />
-            </lvl>
-        '''
+        xml = '<lvl><numFmt val="decimal" /></lvl>'
         level = self._load_from_xml(xml)
-        self.assertEqual(level.num_format, "decimal")
+        self.assertEqual(level.num_format, 'decimal')
 
     def test_restart_attribute(self):
-        xml = b'''
-            <lvl>
-                <lvlRestart val="1" />
-            </lvl>
-        '''
+        xml = '<lvl><lvlRestart val="1" /></lvl>'
         level = self._load_from_xml(xml)
-        self.assertEqual(level.restart, "1")
+        self.assertEqual(level.restart, '1')
 
     def test_associated_paragraph_style_attribute(self):
-        xml = b'''
-            <lvl>
-                <pStyle val="normal" />
-            </lvl>
-        '''
+        xml = '<lvl><pStyle val="normal" /></lvl>'
         level = self._load_from_xml(xml)
-        self.assertEqual(level.paragraph_style, "normal")
+        self.assertEqual(level.paragraph_style, 'normal')
 
     def test_run_properties_child(self):
-        xml = b'''
-            <lvl>
-                <rPr />
-            </lvl>
-        '''
+        xml = '<lvl><rPr /></lvl>'
         level = self._load_from_xml(xml)
-        assert isinstance(level.run_properties, RunProperties), level.run_properties  # noqa
+        assert isinstance(level.run_properties, RunProperties), level.run_properties
 
     def test_paragraph_properties_child(self):
-        xml = b'''
-            <lvl>
-                <pPr />
-            </lvl>
-        '''
+        xml = '<lvl><pPr /></lvl>'
         level = self._load_from_xml(xml)
-        assert isinstance(level.paragraph_properties, ParagraphProperties), level.paragraph_properties  # noqa
+        properties = level.paragraph_properties
+        assert isinstance(properties, ParagraphProperties), properties
+
+    def test_format_is_none_when_not_set(self):
+        xml = '<lvl></lvl>'
+        level = self._load_from_xml(xml)
+        assert level.format_is_none()
+
+    def test_format_is_none_when_set_to_none(self):
+        xml = '<lvl><numFmt val="none" /></lvl>'
+        level = self._load_from_xml(xml)
+        assert level.format_is_none()
+
+    def test_format_is_none_when_set_to_none_case_insensitive(self):
+        xml = '<lvl><numFmt val="NoNe" /></lvl>'
+        level = self._load_from_xml(xml)
+        assert level.format_is_none()