+++ CHANGELOG.rst
@@ -2,6 +2,11 @@
 
 - Decimal font sizes are now handled properly
 
+**0.8.2**
+
+- Paragraphs that have numbering definitions with a level number format of None
+  are no longer considered list items.
+
 **0.8.1**
 
 - Headings in lists no longer break numbering. By default, in the HTML